+++ lib/spout/actions.rb
@@ -2,6 +2,7 @@ module Spout
   class Actions
 
     def interpret(argv)
+      # case argv.first.to_s.scan(/\w/).first
       case argv.first
       when 'new', 'n', 'ne', '-new', '-n', '-ne'
         new_template_dictionary(argv)
@@ -9,8 +10,8 @@ module Spout
         puts "Spout #{Spout::VERSION::STRING}"
       when 'test', 't', 'te', 'tes', '--test', '-t', '-te', '-tes'
         system "bundle exec rake HIDE_PASSING_TESTS=true"
-      when 'tv'
-        system "bundle exec rake"
+      # when 'tv'
+      #   system "bundle exec rake"
       when 'import', 'i', 'im', 'imp', '--import', '-i', '-im', '-imp'
         import_from_csv(argv)
       when 'import_domain', '--import_domain', 'import_domains', '--import_domains'
@@ -25,8 +26,8 @@ module Spout
         generate_charts_and_tables(argv.last(argv.size - 1))
       when 'outliers', '-outliers', '--outliers', 'o', '-o'
         outliers_report(argv)
-      else
-        help
+      # else
+      #   help
       end
     end
 
@@ -52,38 +53,38 @@ EOT
         end
       end
 
-      def help
-        help_message = <<-EOT
+#       def help
+#         help_message = <<-EOT
 
-Usage: spout COMMAND [ARGS]
+# Usage: spout COMMAND [ARGS]
 
-The most common spout commands are:
-  [n]ew             Create a new Spout dictionary.
-                    `spout new <project_name>` creates a new
-                    data dictionary in `./<project_name>`
-  [t]est            Run tests and show failing tests
-  [tv]              Run the tests and show passing and failing
-                    tests
-  [i]mport          Import a CSV file into the JSON dictionary
-  [e]xport [1.0.0]  Export the JSON dictionary to CSV format
-  [c]overage        Coverage report, requires dataset CSVs
-                    in `<project_name>/csvs/<version>`
-  [o]utliers        Outlier report, requires dataset CSVs
-                    in `<project_name>/csvs/<version>`
-  [p]ngs            Generates images for each variable in a
-                    dataset and places them
-                    in `<project_name>/images/<version>/`
-  [g]raphs          Generates JSON graphs for each variable
-                    in a dataset and places them
-                    in `<project_name>/graphs/<version>/`
-  [v]ersion         Returns the version of Spout
+# The most common spout commands are:
+#   [n]ew             Create a new Spout dictionary.
+#                     `spout new <project_name>` creates a new
+#                     data dictionary in `./<project_name>`
+#   [t]est            Run tests and show failing tests
+#   [tv]              Run the tests and show passing and failing
+#                     tests
+#   [i]mport          Import a CSV file into the JSON dictionary
+#   [e]xport [1.0.0]  Export the JSON dictionary to CSV format
+#   [c]overage        Coverage report, requires dataset CSVs
+#                     in `<project_name>/csvs/<version>`
+#   [o]utliers        Outlier report, requires dataset CSVs
+#                     in `<project_name>/csvs/<version>`
+#   [p]ngs            Generates images for each variable in a
+#                     dataset and places them
+#                     in `<project_name>/images/<version>/`
+#   [g]raphs          Generates JSON graphs for each variable
+#                     in a dataset and places them
+#                     in `<project_name>/graphs/<version>/`
+#   [v]ersion         Returns the version of Spout
 
-Commands can be referenced by the first letter:
-  Ex: `spout t`, for test
+# Commands can be referenced by the first letter:
+#   Ex: `spout t`, for test
 
-EOT
-        puts help_message
-      end
+# EOT
+#         puts help_message
+#       end
 
       def new_data_dictionary_export(argv)
         version = argv[1].to_s.gsub(/[^a-zA-Z0-9\.-]/, '_').strip
@@ -91,42 +92,47 @@ EOT
         system "bundle exec rake spout:create #{version_string}"
       end
 
-      def new_template_dictionary(argv)
-        @full_path = File.join(argv[1].to_s.strip)
-        usage = <<-EOT
+#       def new_template_dictionary(argv)
+#         skip_gemfile = (argv.delete('--skip-gemfile') != nil)
+#         @full_path = File.join(argv[1].to_s.strip)
+#         usage = <<-EOT
 
-Usage: spout new FOLDER
+# Usage: spout new FOLDER
 
-The FOLDER must be empty or new.
+# The FOLDER must be empty or new.
 
-EOT
+# EOT
 
-        if @full_path == '' or ( Dir.exists?(@full_path) and (Dir.entries(@full_path) & ['.gitignore', '.ruby-version', '.travis.yml', 'Gemfile', 'Rakefile', 'domains', 'variables', 'test']).size > 0 )
-          puts usage
-          exit(0)
-        end
+#         if @full_path == '' or ( Dir.exists?(@full_path) and (Dir.entries(@full_path) & ['.gitignore', '.ruby-version', '.travis.yml', 'Gemfile', 'Rakefile', 'domains', 'variables', 'test']).size > 0 )
+#           puts usage
+#           exit(0)
+#         end
 
-        FileUtils.mkpath(@full_path)
+#         FileUtils.mkpath(@full_path)
 
-        copy_file 'gitignore', '.gitignore'
-        copy_file 'ruby-version', '.ruby-version'
-        copy_file 'travis.yml', '.travis.yml'
-        copy_file 'Gemfile'
-        copy_file 'Rakefile'
-        directory 'domains'
-        copy_file 'keep', 'domains/.keep'
-        directory 'variables'
-        copy_file 'keep', 'variables/.keep'
-        directory 'test'
-        copy_file 'test/dictionary_test.rb'
-        copy_file 'test/test_helper.rb'
-        puts "         run".colorize( :green ) + "  bundle install".colorize( :light_cyan )
-        Dir.chdir(@full_path)
-        system "bundle install"
-      end
+#         copy_file 'gitignore', '.gitignore'
+#         copy_file 'ruby-version', '.ruby-version'
+#         copy_file 'travis.yml', '.travis.yml'
+#         copy_file 'Gemfile'
+#         copy_file 'Rakefile'
+#         directory 'domains'
+#         copy_file 'keep', 'domains/.keep'
+#         directory 'variables'
+#         copy_file 'keep', 'variables/.keep'
+#         directory 'test'
+#         copy_file 'test/dictionary_test.rb'
+#         copy_file 'test/test_helper.rb'
+#         unless skip_gemfile
+#           puts "         run".colorize( :green ) + "  bundle install".colorize( :light_cyan )
+#           Dir.chdir(@full_path)
+#           system "bundle install"
+#         end
+#       end
 
       def coverage_report(argv)
-        system "bundle exec rake spout:coverage"
+        require 'spout/commands/coverage'
+        Spout::Commands::Coverage.new(standard_version, argv)
+        # system "bundle exec rake spout:coverage"
       end
 
       def outliers_report(argv)
@@ -152,21 +158,21 @@ EOT
         system "bundle exec rake spout:json variables=#{variables.join(',')}"
       end
 
-    private
+    # private
 
-      def copy_file(template_file, file_name = '')
-        file_name = template_file if file_name == ''
-        file_path = File.join(@full_path, file_name)
-        template_file_path = File.join(File.expand_path(File.dirname(__FILE__)), "templates", template_file)
-        puts "      create".colorize( :green ) + "  #{file_name}"
-        FileUtils.copy(template_file_path, file_path)
-      end
+    #   def copy_file(template_file, file_name = '')
+    #     file_name = template_file if file_name == ''
+    #     file_path = File.join(@full_path, file_name)
+    #     template_file_path = File.join(File.expand_path(File.dirname(__FILE__)), "templates", template_file)
+    #     puts "      create".colorize( :green ) + "  #{file_name}"
+    #     FileUtils.copy(template_file_path, file_path)
+    #   end
 
-      def directory(directory_name)
-        directory_path = File.join(@full_path, directory_name)
-        puts "      create".colorize( :green ) + "  #{directory_name}"
-        FileUtils.mkpath(directory_path)
-      end
+    #   def directory(directory_name)
+    #     directory_path = File.join(@full_path, directory_name)
+    #     puts "      create".colorize( :green ) + "  #{directory_name}"
+    #     FileUtils.mkpath(directory_path)
+    #   end
 
   end
 end